hubble/metrics: Introduce v2 HTTP metrics
http v2 metrics are based on L7 response flows rather than both request
and response flows. There was previously some overlap between
http_requests_total and http_responses_total, with the latter having the
status codes. In practice, we should only need one of these, but it
needs the request status code. To achieve this, we now only use the
response flows to count requests, and we invert the source/destination
of the response flows to get the source/destination from the perspective
of the request, rather than the response.

We already have http_request_duration_seconds, which is based on
responses already, but this also leads to the source/destination labels
being "backwards". So like we do with the new http_requests_total
metric, we invert the source and destination.

In addition to consolidating, this is consistent with how http metrics
are exported in the wider ecosystem.

For backwards compatibility, we keep the existing behavior by default,
and expose the new v2 metrics with the same metric names as before using
a feature flag in the metrics context. HTTP v2 metrics can be enabled by
specifying `httpV2` in the `hubble.metric.enabled` helm value:

```
hubble:
  metrics:
    enabled: [ "httpV2" ]
```

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>