hubble/metrics: Use less expensive check for app labels
ParseLabelArrayFromArray can be expensive, so instead, use a more
optimized code path that uses strings.HasPrefix to check if the label
key is correct, then only calls ParseLabel after.

Nested loops are often bad, but in this case, the inner loop is a fixed
size, and should be quite fast. We can't use a map because we want to
use specific app labels before other app labels if they are present and
non-empty.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>